Assignment 7 start of Logic
I added a formalization for the traces in the code further above, as well as in the grammar

The logic is still not completely working for the quantifier raising, so I'm hoping someone can help me with that.

Generating a second version of the tree is a work in progress.